al eliminar el perfil de un senderista, se redirige a home. Modificado el botón de cerrar sesión del nav bar, ahora muestra un popup para confirmar y redirige a home en caso de cerrar sesion

+++ frontend/src/app/common/navbar/navbar.component.ts
@@ -2,6 +2,8 @@
 import { Component, OnInit } from '@angular/core';
 import { Iuser } from '../../../app/modelos/Iuser';
 import { user } from '../../../app/modelos/user';
+import Swal from 'sweetalert2';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-navbar',
@@ -13,7 +15,7 @@ export class NavbarComponent implements OnInit {
   id = '6075a29951822b2eb1f8f43c';
   user: Iuser;
 
-  constructor() { }
+  constructor(private router: Router) { }
 
   ngOnInit(): void {
     this.user = JSON.parse(localStorage.getItem("USER"));
@@ -25,9 +27,39 @@ export class NavbarComponent implements OnInit {
     return us.nickname;
   }
 
-  cerrarSesion(){
-    let us :Iuser = new user("", "", "USUARIO");
-    localStorage.setItem("USER", JSON.stringify(us));
+  botonCerrarSesion(){
+    Swal.fire({
+      title: '¿Seguro que desea cerrar sesión?',
+      showDenyButton: true,
+      showCancelButton: false,
+      confirmButtonText: `Cerrar sesión`,
+      denyButtonText: `Cancelar`,
+      focusDeny:true,
+    }).then((result) => {
+      /* Read more about isConfirmed, isDenied below */
+      if (result.isConfirmed) {
+        try {
+            let us :Iuser = new user("", "", "USUARIO");
+            localStorage.setItem("USER", JSON.stringify(us));
+            this.router.navigate(['/ruta']);
+            
+            let currentUrl = this.router.url;
+            this.router.navigateByUrl('/', {skipLocationChange: true}).then(() => {
+                this.router.navigate([currentUrl]);
+            });
+            
+            Swal.fire('Ha cerrado sesión', '', 'success');
+            // Redirige a la main page
+
+        } catch (error) {
+          Swal.fire({
+            icon: 'error',
+            title: 'Oops...',
+            text: 'No se ha podido completar la operación!',
+          })
+        }
+      }
+    });
   }
 
 }
